+++ l10n_ro_etransport_uit_actions_batch/models/__init__.py
@@ -1,0 +1,6 @@
+# ©  2026 Deltatech
+#              Dorin Hongu <dhongu(@)gmail(.)com
+# See README.rst file on addons root folder for license details
+
+from . import etransport_event
+from . import stock_picking_batch